% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_count.R
\name{na_count}
\alias{na_count}
\title{Count Missing and \% of Total}
\usage{
na_count(x)
}
\arguments{
\item{x}{Vector}
}
\value{
N (\%) of Number missing
}
\description{
Count Missing and \% of Total
}
\examples{
x = 1:100
ind = sample(1:100, 10, replace = F)
x[ind] = NA
na_count(x)

x = 1:200
x[ind] = NA
na_count(x)

}
