% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_comb_iqr.R
\name{str_comb_IQR}
\alias{str_comb_IQR}
\title{Create center with spread data summary for table}
\usage{
str_comb_IQR(x, fun = median, digits = 2, delim = ",", lower = 0.25,
  upper = 0.75)
}
\arguments{
\item{x}{Numeric Vector}

\item{fun}{Function for Central Measure (Defaults to median)}

\item{digits}{number of digits to round}

\item{delim}{delimiter of choice}

\item{lower}{Lower Limit (defaults to Q1) for str_comb_iqr only}

\item{upper}{Upper Limit (defaults to Q3) for str_comb_iqr only}
}
\value{
Character String of the form \code{fun(x) <delim> (Q1,Q3)}
}
\description{
Create center with spread data summary for table
}
\examples{
x = seq(1,100)
str_comb_IQR(x,digits = 3)
str_comb_IQR(x, lower = 0.1, upper = 0.85)
str_comb_IQR(x^2,fun = median, delim = " - ")
str_comb_IQR(x^2,fun = mean, delim = "_")
}
