% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_lessthan.R
\name{to_lessthan}
\alias{to_lessthan}
\title{Function for converting p values to '< 0.01' (or desired power of 10)}
\usage{
to_lessthan(x, digits = 2)
}
\arguments{
\item{x}{Numeric vector}

\item{digits}{desired number of digits to round to}
}
\value{
character vector with values below rounding limit coerced to `< x`
}
\description{
Function for converting p values to '< 0.01' (or desired power of 10)
}
\examples{
TBD

}
